setup internal log package
