+++ exporter/elasticsearchexporter/model.go
@@ -8,32 +8,29 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"hash"
 	"hash/fnv"
 	"math"
 	"slices"
 	"time"
 
-	jsoniter "github.com/json-iterator/go"
 	"go.opentelemetry.io/collector/pdata/pcommon"
-	"go.opentelemetry.io/collector/pdata/plog"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.opentelemetry.io/collector/pdata/ptrace"
-	semconv "go.opentelemetry.io/collector/semconv/v1.22.0"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/ecs"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/encoder"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/exphistogram"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/log"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/mapping"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/objmodel"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/elasticsearchexporter/internal/otel"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/traceutil"
 )
 
-var ErrInvalidTypeForBodyMapMode = errors.New("invalid log record body type for 'bodymap' mapping mode")
-
 type mappingModel interface {
-	encodeLog(pcommon.Resource, string, plog.LogRecord, pcommon.InstrumentationScope, string) ([]byte, error)
+	log.Encoder
+
 	encodeSpan(pcommon.Resource, string, ptrace.Span, pcommon.InstrumentationScope, string) ([]byte, error)
 	encodeSpanEvent(resource pcommon.Resource, resourceSchemaURL string, span ptrace.Span, spanEvent ptrace.SpanEvent, scope pcommon.InstrumentationScope, scopeSchemaURL string) *objmodel.Document
 	upsertMetricDataPointValue(map[uint32]objmodel.Document, pcommon.Resource, string, pcommon.InstrumentationScope, string, pmetric.Metric, dataPoint) error
@@ -49,6 +46,17 @@ type mappingModel interface {
 type encodeModel struct {
 	dedot bool
 	mode  mapping.Mode
+
+	log.Encoder
+}
+
+func newEncodeModel(dedot bool, mode mapping.Mode) *encodeModel {
+	return &encodeModel{
+		dedot: dedot,
+		mode:  mode,
+
+		Encoder: log.New(dedot, mode),
+	}
 }
 
 type dataPoint interface {
@@ -66,170 +74,6 @@ const (
 	spanIDField    = "spanID"
 	attributeField = "attribute"
 )
-
-func (m *encodeModel) encodeLog(resource pcommon.Resource, resourceSchemaURL string, record plog.LogRecord, scope pcommon.InstrumentationScope, scopeSchemaURL string) ([]byte, error) {
-	var document objmodel.Document
-	switch m.mode {
-	case mapping.ModeECS:
-		document = m.encodeLogECSMode(resource, record, scope)
-	case mapping.ModeOTel:
-		document = m.encodeLogOTelMode(resource, resourceSchemaURL, record, scope, scopeSchemaURL)
-	case mapping.ModeBodyMap:
-		return m.encodeLogBodyMapMode(record)
-	default:
-		document = m.encodeLogDefaultMode(resource, record, scope)
-	}
-	// For OTel mode, prefix conflicts are not a problem as otel-data has subobjects: false
-	document.Dedup(m.mode != mapping.ModeOTel)
-
-	var buf bytes.Buffer
-	err := document.Serialize(&buf, m.dedot, m.mode == mapping.ModeOTel)
-	return buf.Bytes(), err
-}
-
-func (m *encodeModel) encodeLogDefaultMode(resource pcommon.Resource, record plog.LogRecord, scope pcommon.InstrumentationScope) objmodel.Document {
-	var document objmodel.Document
-
-	docTimeStamp := record.Timestamp()
-	if docTimeStamp.AsTime().UnixNano() == 0 {
-		docTimeStamp = record.ObservedTimestamp()
-	}
-	document.AddTimestamp("@timestamp", docTimeStamp) // We use @timestamp in order to ensure that we can index if the default data stream logs template is used.
-	document.AddTraceID("TraceId", record.TraceID())
-	document.AddSpanID("SpanId", record.SpanID())
-	document.AddInt("TraceFlags", int64(record.Flags()))
-	document.AddString("SeverityText", record.SeverityText())
-	document.AddInt("SeverityNumber", int64(record.SeverityNumber()))
-	document.AddAttribute("Body", record.Body())
-	m.encodeAttributes(&document, record.Attributes())
-	document.AddAttributes("Resource", resource.Attributes())
-	document.AddAttributes("Scope", scopeToAttributes(scope))
-
-	return document
-}
-
-func (m *encodeModel) encodeLogBodyMapMode(record plog.LogRecord) ([]byte, error) {
-	body := record.Body()
-	if body.Type() != pcommon.ValueTypeMap {
-		return nil, fmt.Errorf("%w: %q", ErrInvalidTypeForBodyMapMode, body.Type())
-	}
-
-	return jsoniter.Marshal(body.Map().AsRaw())
-}
-
-func (m *encodeModel) encodeLogOTelMode(resource pcommon.Resource, resourceSchemaURL string, record plog.LogRecord, scope pcommon.InstrumentationScope, scopeSchemaURL string) objmodel.Document {
-	var document objmodel.Document
-
-	docTimeStamp := record.Timestamp()
-	if docTimeStamp.AsTime().UnixNano() == 0 {
-		docTimeStamp = record.ObservedTimestamp()
-	}
-
-	document.AddTimestamp("@timestamp", docTimeStamp)
-	document.AddTimestamp("observed_timestamp", record.ObservedTimestamp())
-
-	document.AddTraceID("trace_id", record.TraceID())
-	document.AddSpanID("span_id", record.SpanID())
-	document.AddString("severity_text", record.SeverityText())
-	document.AddInt("severity_number", int64(record.SeverityNumber()))
-	document.AddInt("dropped_attributes_count", int64(record.DroppedAttributesCount()))
-
-	otel.EncodeAttributes(&document, record.Attributes())
-	otel.EncodeResource(&document, resource, resourceSchemaURL)
-	otel.EncodeScope(&document, scope, scopeSchemaURL)
-
-	// Body
-	setOTelLogBody(&document, record.Body(), record.Attributes())
-
-	return document
-}
-
-func setOTelLogBody(doc *objmodel.Document, body pcommon.Value, attributes pcommon.Map) {
-	// Determine if this log record is an event, as they are mapped differently
-	// https://github.com/open-telemetry/semantic-conventions/blob/main/docs/general/events.md
-	_, isEvent := attributes.Get("event.name")
-
-	switch body.Type() {
-	case pcommon.ValueTypeMap:
-		if isEvent {
-			doc.AddAttribute("body.structured", body)
-		} else {
-			doc.AddAttribute("body.flattened", body)
-		}
-	case pcommon.ValueTypeSlice:
-		// output must be an array of objects due to ES limitations
-		// otherwise, wrap the array in an object
-		s := body.Slice()
-		allMaps := true
-		for i := 0; i < s.Len(); i++ {
-			if s.At(i).Type() != pcommon.ValueTypeMap {
-				allMaps = false
-			}
-		}
-
-		var outVal pcommon.Value
-		if allMaps {
-			outVal = body
-		} else {
-			vm := pcommon.NewValueMap()
-			m := vm.SetEmptyMap()
-			body.Slice().CopyTo(m.PutEmptySlice("value"))
-			outVal = vm
-		}
-
-		if isEvent {
-			doc.AddAttribute("body.structured", outVal)
-		} else {
-			doc.AddAttribute("body.flattened", outVal)
-		}
-	case pcommon.ValueTypeStr:
-		doc.AddString("body.text", body.Str())
-	default:
-		doc.AddString("body.text", body.AsString())
-	}
-}
-
-func (m *encodeModel) encodeLogECSMode(resource pcommon.Resource, record plog.LogRecord, scope pcommon.InstrumentationScope) objmodel.Document {
-	var document objmodel.Document
-
-	// First, try to map resource-level attributes to ECS fields.
-	ecs.EncodeAttributes(&document, resource.Attributes(), nil)
-
-	// Then, try to map scope-level attributes to ECS fields.
-	scopeAttrsConversionMap := map[string]string{
-		// None at the moment
-	}
-	ecs.EncodeAttributes(&document, scope.Attributes(), scopeAttrsConversionMap)
-
-	// Finally, try to map record-level attributes to ECS fields.
-	recordAttrsConversionMap := map[string]string{
-		"event.name":                         "event.action",
-		semconv.AttributeExceptionMessage:    "error.message",
-		semconv.AttributeExceptionStacktrace: "error.stacktrace",
-		semconv.AttributeExceptionType:       "error.type",
-		semconv.AttributeExceptionEscaped:    "event.error.exception.handled",
-	}
-	ecs.EncodeAttributes(&document, record.Attributes(), recordAttrsConversionMap)
-
-	// Handle special cases.
-	encodeLogAgentNameECSMode(&document, resource)
-	encodeLogAgentVersionECSMode(&document, resource)
-	encodeLogHostOsTypeECSMode(&document, resource)
-	encodeLogTimestampECSMode(&document, record)
-	document.AddTraceID("trace.id", record.TraceID())
-	document.AddSpanID("span.id", record.SpanID())
-	if n := record.SeverityNumber(); n != plog.SeverityNumberUnspecified {
-		document.AddInt("event.severity", int64(record.SeverityNumber()))
-	}
-
-	document.AddString("log.level", record.SeverityText())
-
-	if record.Body().Type() == pcommon.ValueTypeStr {
-		document.AddAttribute("message", record.Body())
-	}
-
-	return document
-}
 
 func (m *encodeModel) encodeDocument(document objmodel.Document) ([]byte, error) {
 	// For OTel mode, prefix conflicts are not a problem as otel-data has subobjects: false
@@ -608,11 +452,11 @@ func (m *encodeModel) encodeSpanDefaultMode(resource pcommon.Resource, span ptra
 	document.AddInt("TraceStatus", int64(span.Status().Code()))
 	document.AddString("TraceStatusDescription", span.Status().Message())
 	document.AddString("Link", spanLinksToString(span.Links()))
-	m.encodeAttributes(&document, span.Attributes())
+	encoder.EncodeAttributes(m.mode, &document, span.Attributes())
 	document.AddAttributes("Resource", resource.Attributes())
 	m.encodeEvents(&document, span.Events())
 	document.AddInt("Duration", durationAsMicroseconds(span.StartTimestamp().AsTime(), span.EndTimestamp().AsTime())) // unit is microseconds
-	document.AddAttributes("Scope", scopeToAttributes(scope))
+	document.AddAttributes("Scope", encoder.ScopeToAttributes(scope))
 	return document
 }
 
@@ -634,14 +478,6 @@ func (m *encodeModel) encodeSpanEvent(resource pcommon.Resource, resourceSchemaU
 	otel.EncodeScope(&document, scope, scopeSchemaURL)
 
 	return &document
-}
-
-func (m *encodeModel) encodeAttributes(document *objmodel.Document, attributes pcommon.Map) {
-	key := "Attributes"
-	if m.mode == mapping.ModeRaw {
-		key = ""
-	}
-	document.AddAttributes(key, attributes)
 }
 
 func (m *encodeModel) encodeEvents(document *objmodel.Document, events ptrace.SpanEventSlice) {
@@ -670,104 +506,6 @@ func spanLinksToString(spanLinkSlice ptrace.SpanLinkSlice) string {
 // which is the format the Duration field is stored in the Span.
 func durationAsMicroseconds(start, end time.Time) int64 {
 	return (end.UnixNano() - start.UnixNano()) / 1000
-}
-
-func scopeToAttributes(scope pcommon.InstrumentationScope) pcommon.Map {
-	attrs := pcommon.NewMap()
-	attrs.PutStr("name", scope.Name())
-	attrs.PutStr("version", scope.Version())
-	for k, v := range scope.Attributes().AsRaw() {
-		attrs.PutStr(k, v.(string))
-	}
-	return attrs
-}
-
-func encodeLogAgentNameECSMode(document *objmodel.Document, resource pcommon.Resource) {
-	// Parse out telemetry SDK name, language, and distro name from resource
-	// attributes, setting defaults as needed.
-	telemetrySdkName := "otlp"
-	var telemetrySdkLanguage, telemetryDistroName string
-
-	attrs := resource.Attributes()
-	if v, exists := attrs.Get(semconv.AttributeTelemetrySDKName); exists {
-		telemetrySdkName = v.Str()
-	}
-	if v, exists := attrs.Get(semconv.AttributeTelemetrySDKLanguage); exists {
-		telemetrySdkLanguage = v.Str()
-	}
-	if v, exists := attrs.Get(semconv.AttributeTelemetryDistroName); exists {
-		telemetryDistroName = v.Str()
-		if telemetrySdkLanguage == "" {
-			telemetrySdkLanguage = "unknown"
-		}
-	}
-
-	// Construct agent name from telemetry SDK name, language, and distro name.
-	agentName := telemetrySdkName
-	if telemetryDistroName != "" {
-		agentName = fmt.Sprintf("%s/%s/%s", agentName, telemetrySdkLanguage, telemetryDistroName)
-	} else if telemetrySdkLanguage != "" {
-		agentName = fmt.Sprintf("%s/%s", agentName, telemetrySdkLanguage)
-	}
-
-	// Set agent name in document.
-	document.AddString("agent.name", agentName)
-}
-
-func encodeLogAgentVersionECSMode(document *objmodel.Document, resource pcommon.Resource) {
-	attrs := resource.Attributes()
-
-	if telemetryDistroVersion, exists := attrs.Get(semconv.AttributeTelemetryDistroVersion); exists {
-		document.AddString("agent.version", telemetryDistroVersion.Str())
-		return
-	}
-
-	if telemetrySdkVersion, exists := attrs.Get(semconv.AttributeTelemetrySDKVersion); exists {
-		document.AddString("agent.version", telemetrySdkVersion.Str())
-		return
-	}
-}
-
-func encodeLogHostOsTypeECSMode(document *objmodel.Document, resource pcommon.Resource) {
-	// https://www.elastic.co/guide/en/ecs/current/ecs-os.html#field-os-type:
-	//
-	// "One of these following values should be used (lowercase): linux, macos, unix, windows.
-	// If the OS you’re dealing with is not in the list, the field should not be populated."
-
-	var ecsHostOsType string
-	if semConvOsType, exists := resource.Attributes().Get(semconv.AttributeOSType); exists {
-		switch semConvOsType.Str() {
-		case "windows", "linux":
-			ecsHostOsType = semConvOsType.Str()
-		case "darwin":
-			ecsHostOsType = "macos"
-		case "aix", "hpux", "solaris":
-			ecsHostOsType = "unix"
-		}
-	}
-
-	if semConvOsName, exists := resource.Attributes().Get(semconv.AttributeOSName); exists {
-		switch semConvOsName.Str() {
-		case "Android":
-			ecsHostOsType = "android"
-		case "iOS":
-			ecsHostOsType = "ios"
-		}
-	}
-
-	if ecsHostOsType == "" {
-		return
-	}
-	document.AddString("host.os.type", ecsHostOsType)
-}
-
-func encodeLogTimestampECSMode(document *objmodel.Document, record plog.LogRecord) {
-	if record.Timestamp() != 0 {
-		document.AddTimestamp("@timestamp", record.Timestamp())
-		return
-	}
-
-	document.AddTimestamp("@timestamp", record.ObservedTimestamp())
 }
 
 // TODO use https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/internal/exp/metrics/identity
